fa_eval/requirements.txt
mmif-python
pyannote.metrics
pandas
